Ensure the AppStream data can be used for the repo's appstream branch
Flatpak expects a file like <id>.appdata.xml to be provided by the app, which
is not the case for GNOME Weather, so get it renamed when creating the bundle.

This should be reverted once the file gets renamed upstream.
See bug https://bugzilla.gnome.org/show_bug.cgi?id=777052