Revert "Ensure the AppStream data can be used for the repo's appstream branch"
After some more consideration and tests on this, I'm not sure this is actually
correct, as org.gnome.Weather.Application since to be the ID used everywhere,
from the desktop file to the D-Bus name, Exec line in desktop file, and so on.

Thus, I'm reverting this patch now and will comment in the issue upstream
at https://bugzilla.gnome.org/show_bug.cgi?id=777052. Sorry for the confusion.

This reverts commit 206a5ad7845d9dfa4b4b9364ea44971b66dfc6cb.