gnome-games: Update the stable version of GNOME Games, to make it usable
The current stable flatpak for GNOME Games crashes too often when several
ROMs are available while fetching the covers, which seems to be improved
by building against commit d1cde2b3 (the first code-related commit after
the stable 3.22 release), so we'd better use that.

Besides, we make the most of this commit to update the stable versions
of all the related dependencies with the exception of tracker, whose
latest version does not build against the 3.22 runtime.