gnome-games: Used unreleased version of grilo to fix crash fetching covers
It turns out that the previous commit was wrong and that just using a
slightly more recent version of GNOME Games was not enough to fix the
crash observed while fetching covers. Instead, we need to pull in an
unreleased version of grilo (from a commit which includes the very first
bugfix right after the latest stable release, so it should be fine).

This patch restores GNOME Games source to use the stable tarball and
updates the version of grilo to include the needed patch. Any other
dependency updated before (to stable versions) are kept.

See grilo bug: https://bugzilla.gnome.org/show_bug.cgi?id=771338