Rhythmbox: Change the runtime from master to 3.28
Gnome-sdk master builds gstreamer from git, which means depricated
APIs are disabled. Rhythmbox depends on a depricated gst API which
makes it unbuildable with the master runtime.

Since there are does not seem to be the will to replace the
depricated api currently, switch the runtime so at least the thing
is building.

https://gitlab.gnome.org/GNOME/rhythmbox/issues/1649